osx build: small fix

+++ tools/MacOSX/build-plugins.sh
@@ -165,12 +165,14 @@ for plugin in "$PLUGINDIR/OFX/Natron"/*.ofx.bundle; do
     # relative path to the Natron.app/Contents/Frameworks dir
     frameworks=../../../../../../Frameworks
     for lib in lib*.dylib; do
-        if [ -f "$frameworks/$lib" ]; then
-            rm "$lib"
-        else
-            mv "$lib" "$frameworks/$lib"
+        if [ -f "$lib" ]; then
+            if [ -f "$frameworks/$lib" ]; then
+                rm "$lib"
+            else
+                mv "$lib" "$frameworks/$lib"
+            fi
+            ln -sf "$frameworks/$lib" "$lib"
         fi
-        ln -sf "$frameworks/$lib" "$lib"
     done
     if [ "$COMPILER" = "gcc" ]; then # use gcc's libraries everywhere
         for l in gcc_s.1 gomp.1 stdc++.6; do
